Simplify .testing; trying to squish AppVeyor bug
