+++ tests/conftest.py
@@ -1,6 +1,28 @@
-from aspen.testing import teardown
-from aspen.testing.pytest_fixtures import harness, fs, processor
-from aspen.testing.pytest_fixtures import sys_path, sys_path_scrubber
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+import sys
+
+import pytest
+from aspen.testing import Harness, teardown
+
+
+@pytest.yield_fixture
+def sys_path_scrubber():
+    before = set(sys.path)
+    yield
+    after = set(sys.path)
+    for name in after - before:
+        sys.path.remove(name)
+
+
+@pytest.yield_fixture
+def harness(sys_path_scrubber):
+    harness = Harness()
+    yield harness
+    harness.teardown()
 
 
 def pytest_runtest_teardown():
